fix merge weirdness
